Started reusing existing ValidationError
