+++ npm-api-maker/src/commands-pool.cjs
@@ -1,11 +1,11 @@
 const Api = require("./api.cjs")
 const CommandSubmitData = require("./command-submit-data.cjs")
 const CustomError = require("./custom-error.cjs")
-const CustomValidationError = require("./custom-validation-error.cjs")
 const Deserializer = require("./deserializer.cjs")
 const {dig} = require("@kaspernj/object-digger")
 const FormDataObjectizer = require("form-data-objectizer")
 const Serializer = require("./serializer.cjs")
+const ValidationError = require("./validation-error.cjs")
 
 module.exports = class ApiMakerCommandsPool {
   static addCommand(data, args = {}) {
@@ -140,7 +140,11 @@ module.exports = class ApiMakerCommandsPool {
           let error
 
           if ("validation_errors" in commandResponseData) {
-            error = new CustomValidationError("Command failed with validation errors", {response: commandResponseData})
+            const validationErrors = new ValidationErrors({
+              model: digg(commandResponseData, "model"),
+              validationErrors: digg(commandResponseData, "validationErrors")
+            })
+            error = new ValidationError(validationErrors(), {response: commandResponseData})
           } else {
             error = new CustomError("Command failed", {response: commandResponseData})
           }
